+++ bin/oe/cli/mod.rs
@@ -748,6 +748,10 @@ usage! {
             "--log-file=[FILENAME]",
             "Specify a filename into which logging should be appended.",
 
+            ARG arg_shutdown_on_missing_block_import: (Option<u64>) = None, or |c: &Config| c.misc.as_ref()?.shutdown_on_missing_block_import.clone(),
+            "--shutdown-on-missing-block-import=[STRING]",
+            "Shuts down if no block has been imported for N seconds. Defaults to None. Set to None or 0 to disable this feature. This setting is only respected by the HBBFT Engine",
+
         ["Footprint Options"]
             FLAG flag_scale_verifiers: (bool) = false, or |c: &Config| c.footprint.as_ref()?.scale_verifiers.clone(),
             "--scale-verifiers",
@@ -1039,6 +1043,8 @@ struct Misc {
     color: Option<bool>,
     ports_shift: Option<u16>,
     unsafe_expose: Option<bool>,
+    /// seconds, until the system shuts down if no block has been produced. None disables this feature.
+    shutdown_on_missing_block_import: Option<u64>,
 }
 
 #[cfg(test)]
@@ -1450,6 +1456,7 @@ mod tests {
                 arg_log_file: Some("/var/log/openethereum.log".into()),
                 flag_no_color: false,
                 flag_no_config: false,
+                arg_shutdown_on_missing_block_import: None,
             }
         );
     }
@@ -1640,6 +1647,7 @@ mod tests {
                     color: Some(true),
                     ports_shift: Some(0),
                     unsafe_expose: Some(false),
+                    shutdown_on_missing_block_import: None,
                 }),
                 stratum: None,
             }