+++ crates/ethcore/src/client/client.rs
@@ -1327,8 +1327,7 @@ impl Client {
                     // if the engine still needs that block, we are not going to prune it.
                     if let Some(protected_block) = self.engine.pruning_protection_block_number() {
                         if earliest_era > protected_block {
-                            info!(target: "pruning", "Detected attempt from pruning ancient block that is still required by the engine. protected block: {protected_block}, earliest_era: {earliest_era}");
-                            //earliest_era = protected_block - 1;
+                            debug!(target: "pruning", "Detected attempt from pruning ancient block that is still required by the engine. protected block: {protected_block}, earliest_era: {earliest_era}");
                             break;
                         }
                     }
@@ -3246,6 +3245,10 @@ impl super::traits::EngineClient for Client {
 
     fn demand_shutdown(&self) {
         self.shutdown.demand_shutdown();
+    }
+
+    fn config_shutdown_on_missing_block_import(&self) -> Option<u64> {
+        self.config.shutdown_on_missing_block_import
     }
 
     fn create_pending_block_at(