+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## Diamond Node Software 3.3.5-hbbft-0.9.3
+
+[Autoshutdown after a period without block import] https://github.com/DMDcoin/diamond-node/issues/78
+
+Those examples show how to confige the node to activate this feature, restarting the node if no block import has been detected for 1800 seconds (30 minutes)
+
+to activate feature via CLI Arg:
+`--shutdown-on-missing-block-import=1800`
+
+or in node.toml
+node.toml:
+```
+[Misc]
+shutdown_on_missing_block_import = 1800
+```
+
 ## Diamond Node Software 3.3.5-hbbft-0.9.2
 
 - [FIXED: pruning as root cause for stage 3 errors] https://github.com/DMDcoin/diamond-node/issues/68